+++ src/dotools_py/pl/__init__.py
@@ -5,4 +5,4 @@ from dotools_py.pl._expression import barplot, boxplot, violinplot
 from dotools_py.pl._heatmap import heatmap, heatmap_foldchange
 from dotools_py.pl._StatsPlotter import StatsPlotter, TestData
 from dotools_py.pl._visium import layers, slides
-from dotools_py.pl._experimental import lineplot
+from dotools_py.pl._experimental import lineplot, ridgeplot